refactored a bit for better reuse
