+++ core/src/main/java/hudson/model/Actionable.java
@@ -134,29 +134,6 @@ public abstract class Actionable extends AbstractModelObject implements ModelObj
     }
 
     public ContextMenu doContextMenu(StaplerRequest request, StaplerResponse response) throws Exception {
-        ContextMenu c = new ContextMenu();
-
-        WebApp webApp = WebApp.getCurrent();
-        final Script s = webApp.getMetaClass(this).getTearOff(JellyClassTearOff.class).findScript("sidepanel");
-        if (s==null) {
-            // fallback
-            c.addAll(getActions());
-        } else {
-            JellyFacet facet = webApp.getFacet(JellyFacet.class);
-            request.setAttribute("taskTags",c);
-            request.setAttribute("mode","side-panel");
-            facet.scriptInvoker.invokeScript(request,response,new Script() {
-                public Script compile() throws JellyException {
-                    return this;
-                }
-
-                public void run(JellyContext context, XMLOutput output) throws JellyTagException {
-                    Functions.initPageVariables(context);
-                    s.run(context,output);
-                }
-            }, this,new XMLOutput(new DefaultHandler()));
-        }
-
-        return c;
+        return new ContextMenu().from(this,request,response);
     }
 }
